Try a hard-coded AMI id
gh-2

When I try to connect with Instance Connect, I get:

There was a problem connecting to your instance
Log in failed. If this instance has just started up, wait a few minutes
and try again. Otherwise, ensure the instance is running on an AMI that
supports EC2 Instance Connect.

The linked page,

https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-instance-connect-methods.html?icmpid=docs_ec2_console#ic-limitations

led eventually (not sure how I got there) to:

https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-instance-connect-set-up.html?icmpid=docs_ec2_console#ec2-instance-connect-setup-security-group

which says:

Task 2: (Conditional) Install EC2 Instance Connect on an instance
You can skip this task if you used one of the following AMIs to launch
your instance because they come preinstalled with EC2 Instance Connect:

Amazon Linux 2 2.0.20190618 or later

Ubuntu 20.04 or later

For earlier versions of these AMIs, you must install Instance Connect on
every instance that will support connecting using Instance Connect.

I think that prior to this commit, I was not using Amazon Linux 2.

I went in the console and launched an instance using the Amazon Linux 2
AMI that I am attempting to set in this commit, and I was able to
connect with Instance Connect to it.

Its AMI name, the one that worked, had like "amzn2-...", whereas the CDK
instance I currently ahve says like "amzn-...".